split out examples into pages, inline config in page
