+++ examples/sandbox/src/Router.js
@@ -1,15 +1,21 @@
 import * as React from "react";
-import { Switch, Route } from "react-router-dom";
-import "./App.css";
-import App from "./App";
-import WorkplaceSearch from "./WorkplaceSearch";
+import { Switch, Route, Redirect } from "react-router-dom";
+import WorkplaceSearch from "./pages/workplace-search";
+import AppSearch from './pages/app-search';
+import Elasticsearch from './pages/elasticsearch'
 
 export default function Router() {
   return (
     <div className="App">
       <Switch>
         <Route exact path="/">
-          <App />
+          <Redirect to="/app-site-search" />
+        </Route>
+        <Route exact path="/app-site-search">
+          <AppSearch />
+        </Route>
+        <Route exact path="/elasticsearch">
+          <Elasticsearch />
         </Route>
         <Route exact path="/workplace-search">
           <WorkplaceSearch />
