+++ examples/sandbox/src/index.js
@@ -1,9 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom";
 import { BrowserRouter } from "react-router-dom";
-import "./index.css";
 import Router from "./Router";
-import * as serviceWorker from "./serviceWorker";
 
 ReactDOM.render(
   <BrowserRouter>
@@ -11,8 +9,3 @@ ReactDOM.render(
   </BrowserRouter>,
   document.getElementById("root")
 );
-
-// If you want your app to work offline and load faster, you can change
-// unregister() to register() below. Note this comes with some pitfalls.
-// Learn more about service workers: http://bit.ly/CRA-PWA
-serviceWorker.unregister();
